moved leaderboard and common components to folders
