+++ frontend/app/src/components/common/__tests__/Button.spec.tsx
@@ -3,10 +3,10 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import nock from 'nock';
 import React from 'react';
-import { setupStore } from '../../__test__/__mockData__/setupStore';
-import { user } from '../../__test__/__mockData__/user';
-import { mockUsehistory } from '../../__test__/__mockFn__/useHistory';
-import Button from './Button';
+import { setupStore } from '../../../__test__/__mockData__/setupStore';
+import { user } from '../../../__test__/__mockData__/user';
+import { mockUsehistory } from '../../../__test__/__mockFn__/useHistory';
+import Button from '../Button';
 
 beforeAll(() => {
   nock.disableNetConnect();
