+++ frontend/app/src/pages/leaderboard/leaderboardItem/__tests__/LeaerboardItem.spec.tsx
@@ -3,10 +3,10 @@ import { render, screen } from '@testing-library/react';
 import userEvent from '@testing-library/user-event';
 import nock from 'nock';
 import React from 'react';
-import { setupStore } from '../../../__test__/__mockData__/setupStore';
-import { user } from '../../../__test__/__mockData__/user';
-import { mockUsehistory } from '../../../__test__/__mockFn__/useHistory';
-import { LeaerboardItem } from './index';
+import { setupStore } from '../../../../__test__/__mockData__/setupStore';
+import { user } from '../../../../__test__/__mockData__/user';
+import { mockUsehistory } from '../../../../__test__/__mockFn__/useHistory';
+import { LeaerboardItem } from '../index';
 
 beforeAll(() => {
   nock.disableNetConnect();
